(Adding missing files)
First part of a major clean-up for XMLVM:
In the coming weeks we will continue to restructure XMLVM to make it easier to develop and maintain certain parts of XMLVM.

We break XMLVM into a "core" part and "plugins". In this first CL we move output processes that belong to a certain plugin to the new plugin directory. In the next weeks we will move other resource, such as XSL stylesheets, libraries etc and create a mechanism that allows for independent development of these plugins.

git-svn-id: svn://svn.code.sf.net/p/xmlvm/code/trunk@2354 0874c2aa-48f6-4a0f-8a6b-e32f4465a257